build.py: quadruple backslash pathnames in re.sub()
Singley and doubley backslashed path names cause an internal error in
the re.sub() function.

According to the Python documentation, regarding the "repl" string arg
for re.sub():

	repl can be a string or a function; if it is a string, any
	backslash escapes in it are processed. \n is converted to a
	single newline character, \r is converted to a carriage return,
	and so forth. Unknown escapes of ASCII letters are reserved for
	future use and treated as errors.

So the single backslashes in paths cause errors (i.e.,
"Citrix\Xenserver" resolves to "\X" which is an invalid escape symbol),
and the double backslashes "Citrix\\Xenserver" are preprocessed to
"Citrix\Xenserver" which is then re-evaluated as an invalid escape
symbol when the substitution is performed (AFAICT).

Quadruple backslashing fixes this issue, in "Citrix\\\\Xenserver" both
sets of "\\" resolve to a single "\", which yields "Citrix\\Xenserver"
and is a properly escaped backslash.

Signed-off-by: Bobby Eshleman <bobbyeshleman@gmail.com>